Are toate cerintele indeplinite cu exceptia dynamic cast, bombele sunt implementate

+++ Game.cpp
@@ -9,7 +9,7 @@
 
     Game::Game(const Game &other): mbr{other.mbr}, m{other.m}, player{other.player}{
         obj.clear();
-        for(int i=0; i<other.obj.size(); i++)
+        for(unsigned long i=0; i<other.obj.size(); i++)
         {
             Objective * obj_cl = other.obj[i]->clone();
             obj.emplace_back(obj_cl);
